simplify jobs and add trade action

+++ server/src/main/java/massim/scenario/city/data/TeamState.java
@@ -23,7 +23,7 @@ public class TeamState {
         return massium;
     }
 
-    public void addMassium(long amount){
+    public void addMassium(int amount){
         massium += amount;
     }
